// src/app/country/model/country.enum.ts
export enum CountryEnum {
    MEXICO = 'MX',
    USA = 'US',
    GREATBRITAIN = 'GB'
}

export enum FormatMexico {
    date = 'DD/MM/YYYY',
    dateTime = 'DD/MM/YYYY, HH:mm',
    time = 'HH:mm'
}

export enum FormatUsa {
    date = 'MM/DD/YYYY',
    dateTime = 'MM/DD/YYYY, HH:mm',
    time = 'HH:mm'
}

export enum FormatEu {
    date = 'DD.MM.YYYY',
    dateTime = 'DD.MM.YYYY HH:mm',
    time = 'HH:mm'
}

export enum Format {
    date = 'date',
    dateTime = 'dateTime',
    time = 'time'
}

export enum BusinessLineCode {
    Readymix = 'RMX',
    Cement = 'CEM'
}